Work with over 20MB files

+++ spit/randdd.c
@@ -180,7 +180,7 @@ int main(int argc, char *argv[]) {
     int fd = open(device, O_DIRECT | O_WRONLY | O_TRUNC | O_EXCL, S_IRUSR | S_IWUSR);
     if (fd >= 0) {
       
-      for (double offset = startpos; offset <= endpos - size; offset += gap) {
+      for (double offset = startpos; offset <= (double) endpos - (double)size; offset += gap) {
 	
 	size_t aloff = alignedNumber((size_t)(offset + 0.5), 4096);
